itechsmart-ninja/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI Providers
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1

# Docker SDK
docker==6.1.3

# Task Queue
celery==5.3.4
redis==5.0.1

# AWS SDK (for S3 deployments)
boto3==1.34.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
# Web scraping and search
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
pillow==10.1.0
replicate==0.22.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0

# VM Management
docker>=6.1.0
psutil>=5.9.0

# Task Scheduling
APScheduler>=3.10.0
croniter>=2.0.0

# MCP Data Sources (Feature 10)
psycopg2-binary>=2.9.0
pymysql>=1.1.0
pymongo>=4.5.0
redis>=5.0.0
elasticsearch>=8.10.0
boto3>=1.28.0
aiohttp>=3.9.0

# Video Generation (Feature 12)
replicate>=0.15.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
